renaming direcotires and creating new order for steps
