+++ app/src/app/app.component.spec.ts
@@ -28,15 +28,15 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatButtonModule } from '@angular/material/button';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ModalModule } from './components/specmath-modal/modal.module';
+import { ModalModule } from './components/merge-modal/modal.module';
 import { queryElement } from 'src/shared/functions';
-import { DisplayResultsModule } from './components/spec-math-display-results/display-results.module';
+import { DisplayResultsModule } from './components/display-results/display-results.module';
 import { YamlRenderModule } from './components/yaml-render/yaml-render.module';
 import { MatListModule } from '@angular/material/list';
 import { DefaultsPageComponent } from './components/defaults-page/defaults-page.component';
 import { AboutPageComponent } from './components/about-page/about-page.component';
 import { Router } from '@angular/router';
-import { DisplayResultsComponent } from './components/spec-math-display-results/display-results.component';
+import { DisplayResultsComponent } from './components/display-results/display-results.component';
 import { EmptyPageComponent } from './components/empty-page/empty-page.component';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { OperationService } from 'src/shared/services/operation.service';
