+++ app/src/app/app.module.ts
@@ -22,8 +22,8 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
-import { ModalModule } from './components/specmath-modal/modal.module';
-import { DisplayResultsModule } from './components/spec-math-display-results/display-results.module';
+import { ModalModule } from './components/merge-modal/modal.module';
+import { DisplayResultsModule } from './components/display-results/display-results.module';
 import { YamlRenderModule } from './components/yaml-render/yaml-render.module';
 import { MatListModule } from '@angular/material/list';
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -31,7 +31,7 @@ import { OperationService } from 'src/shared/services/operation.service';
 import { RouterModule, Routes } from '@angular/router';
 import { AboutPageComponent } from './components/about-page/about-page.component';
 import { DefaultsPageComponent } from './components/defaults-page/defaults-page.component';
-import { DisplayResultsComponent } from './components/spec-math-display-results/display-results.component';
+import { DisplayResultsComponent } from './components/display-results/display-results.component';
 import { EmptyPageComponent } from './components/empty-page/empty-page.component';
 import { SideNavService } from 'src/shared/services/sidenav.service';
 import { routes } from 'src/shared/routes';
