+++ app/src/app/components/display-results/display-results.component.html
