+++ app/src/app/components/display-results/display-results.component.spec.ts
