+++ app/src/app/components/merge-modal/confirm-operation/confirm-operation.component.scss
