+++ app/src/app/components/merge-modal/defaults-file-upload/defaults-file-upload.component.scss
