+++ app/src/app/components/merge-modal/modal.component.scss
