+++ app/src/app/components/merge-modal/modal.component.spec.ts
