+++ app/src/app/components/merge-modal/resolve-conflict/resolve-conflict.component.html
