+++ app/src/app/components/merge-modal/spec-files-upload/spec-files-upload.component.scss
