+++ app/src/app/components/merge-modal/spec-name-input/spec-name-input.component.spec.ts
